# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Arc Stablecoin DApp API — Conditional Escrow, Payroll Vesting, Cross-chain Transfers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /htlc:
    get:
      operationId: listHtlcSwaps
      tags: [htlc]
      summary: List HTLC swaps
      parameters:
        - in: query
          name: address
          schema:
            type: string
          description: Filter by depositor or recipient address
        - in: query
          name: status
          schema:
            type: string
            enum: [active, claimed, refunded]
      responses:
        "200":
          description: List of HTLC swaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HTLCSwap"

    post:
      operationId: createHtlcSwap
      tags: [htlc]
      summary: Record a new HTLC swap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HTLCSwapInput"
      responses:
        "201":
          description: HTLC swap created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTLCSwap"
        "400":
          description: Invalid input
  /htlc/{id}:
    get:
      operationId: getHtlcSwap
      tags: [htlc]
      summary: Get HTLC swap by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: HTLC swap details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTLCSwap"
        "404":
          description: Not found
  /htlc/{id}/claim:
    post:
      operationId: claimHtlcSwap
      tags: [htlc]
      summary: Record a claim on an HTLC swap (reveals preimage)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HTLCClaimInput"
      responses:
        "200":
          description: HTLC claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTLCSwap"
        "404":
          description: Not found
  /htlc/{id}/refund:
    post:
      operationId: refundHtlcSwap
      tags: [htlc]
      summary: Record a refund on an expired HTLC swap
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HTLCRefundInput"
      responses:
        "200":
          description: HTLC refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTLCSwap"
        "404":
          description: Not found
  /htlc/{id}/relay:
    post:
      operationId: relayHtlcSwap
      tags: [htlc]
      summary: Record CCTP attestation relay on destination chain (crosschain_cctp mode only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HTLCRelayInput"
      responses:
        "200":
          description: HTLC relay recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTLCSwap"
        "404":
          description: Not found
...
  /splits:
    get:
      operationId: listSplits
      tags: [split]
      summary: List payment splits
      parameters:
        - in: query
          name: creator
          schema:
            type: string
          description: Filter by creator address
      responses:
        "200":
          description: List of splits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Split"

    post:
      operationId: createSplit
      tags: [split]
      summary: Create a new fixed-share payment split
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitInput"
      responses:
        "201":
          description: Split created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Split"
        "400":
          description: Invalid input
  /splits/{id}:
    get:
      operationId: getSplit
      tags: [split]
      summary: Get split by DB id (includes distribution history)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Split record with distributions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Split"
        "404":
          description: Not found
  /splits/{id}/distribute:
    post:
      operationId: distributeSplit
      tags: [split]
      summary: Record a distribution event for a split
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitDistributeInput"
      responses:
        "200":
          description: Updated split after distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Split"
        "400":
          description: Invalid input
        "404":
          description: Not found
  /splits/{id}/deactivate:
    patch:
      operationId: deactivateSplit
      tags: [split]
      summary: Deactivate a split (creator only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitDeactivateInput"
      responses:
        "200":
          description: Deactivated split
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Split"
        "404":
          description: Not found
components:
  schemas:
...
    HTLCSwap:
      type: object
      required: [id, depositor, recipient, token, amount, hashlock, timelock, status, contractAddress, txHash, chainId, createdAt]
      properties:
        id:
          type: integer
        depositor:
          type: string
          description: Address that locked tokens; can refund after timelock
        recipient:
          type: string
          description: "Single-chain: address that claims here. Crosschain CCTP: for display only — actual recipient is mintRecipient on dest chain."
        token:
          type: string
          enum: [USDC, EURC]
        amount:
          type: string
          description: Token amount in raw 6-decimal base units
        hashlock:
          type: string
          description: "keccak256(abi.encode(preimage)) — computed off-chain by depositor"
        timelock:
          type: integer
          description: Unix timestamp after which depositor can refund
        status:
          type: string
          enum: [active, claimed, relayed, refunded]
        preimage:
          type: ["string", "null"]
          description: Revealed preimage (bytes32 hex); null until claimed
        claimTxHash:
          type: ["string", "null"]
          description: Claim transaction hash (also triggers CCTP burn for crosschain mode)
        refundTxHash:
          type: ["string", "null"]
          description: Refund transaction hash
        contractAddress:
          type: string
        txHash:
          type: string
          description: Creation transaction hash
        onChainId:
          type: ["integer", "null"]
          description: On-chain HTLC ID from the contract
        chainId:
          type: integer
        swapMode:
          type: ["string", "null"]
          enum: [single_chain, crosschain_cctp, null]
          description: "single_chain = tokens stay on Arc; crosschain_cctp = claim triggers depositForBurn via CCTP"
        destinationDomain:
          type: ["integer", "null"]
          description: "CCTP destination domain (0=Eth Sepolia, 3=Arb Sepolia, 6=Base Sepolia). Null for single-chain."
        mintRecipient:
          type: ["string", "null"]
          description: "bytes32 hex of the USDC recipient on the destination chain. Null for single-chain."
        maxFee:
          type: ["string", "null"]
          description: "Circle attestation fee in raw base units. Null for single-chain."
        minFinalityThreshold:
          type: ["integer", "null"]
          description: "CCTP finality threshold: 2000=finalized, 1000=fast. Null for single-chain."
        relayTxHash:
          type: ["string", "null"]
          description: "Tx hash of the CCTP attestation relay on the destination chain. Null until relayed."
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HTLCSwapInput:
      type: object
      required: [depositor, recipient, token, amount, hashlock, timelock, contractAddress, txHash]
      properties:
        depositor:
          type: string
        recipient:
          type: string
        token:
          type: string
          enum: [USDC, EURC]
        amount:
          type: string
        hashlock:
          type: string
          description: keccak256(abi.encode(preimage)) as 0x-prefixed hex bytes32
        timelock:
          type: integer
        contractAddress:
          type: string
        txHash:
          type: string
        chainId:
          type: integer
        onChainId:
          type: integer
        swapMode:
          type: string
          enum: [single_chain, crosschain_cctp]
        destinationDomain:
          type: integer
        mintRecipient:
          type: string
          description: bytes32 hex of recipient on destination chain
        maxFee:
          type: string
        minFinalityThreshold:
          type: integer
    HTLCClaimInput:
      type: object
      required: [txHash, preimage]
      properties:
        txHash:
          type: string
        preimage:
          type: string
          description: The secret bytes32 value that hashes to hashlock
    HTLCRefundInput:
      type: object
      required: [txHash]
      properties:
        txHash:
          type: string
    HTLCRelayInput:
      type: object
      required: [txHash]
      properties:
        txHash:
          type: string
          description: Tx hash of the CCTP attestation relay call on the destination chain
...
    Split:
      type: object
      properties:
        id:
          type: integer
        onChainId:
          type: integer
        creator:
          type: string
        token:
          type: string
          enum: [USDC, EURC]
        recipients:
          type: array
          items:
            type: string
        shares:
          type: array
          items:
            type: integer
          description: Basis points per recipient; must sum to 10000
        description:
          type: string
        totalDistributed:
          type: string
          description: Cumulative raw base units distributed
        active:
          type: boolean
        contractAddress:
          type: string
        txHash:
          type: string
        chainId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SplitInput:
      type: object
      required: [creator, token, recipients, shares, contractAddress, txHash]
      properties:
        creator:
          type: string
          description: Address that created the split
        token:
          type: string
          enum: [USDC, EURC]
        recipients:
          type: array
          items:
            type: string
          description: 2-20 recipient addresses
        shares:
          type: array
          items:
            type: integer
          description: Basis points per recipient; must sum to 10000
        description:
          type: string
        contractAddress:
          type: string
          description: SplitPayment contract address
        txHash:
          type: string
          description: createSplit() transaction hash
        onChainId:
          type: integer
        chainId:
          type: integer
    SplitDistributeInput:
      type: object
      required: [amount, txHash, distributor]
      properties:
        amount:
          type: string
          description: USDC/EURC amount in raw 6-decimal base units
        txHash:
          type: string
          description: distribute() transaction hash
        distributor:
          type: string
          description: Address that called distribute
    SplitDeactivateInput:
      type: object
      required: [txHash]
      properties:
        txHash:
          type: string